Accelerate anc4 to 95% by 2030

Accelerate anc4 by aiming at fixed target of 95% by 2030 for countries with at least 2 reported values. For others, it is business as usual.

Usage

accelerate_target_anc4(
  df,
  scenario_col = "scenario",
  default_scenario = "default",
  bau_scenario = "historical",
  scenario_name = "acceleration_target",
  ind_ids = billion_ind_codes("uhc"),
  ...
)

Arguments

df

Data frame in long format, where 1 row corresponds to a specific country, year, and indicator.

scenario_col

Column name of column with scenario identifiers. Useful for calculating contributions on data in long format rather than wide format.

default_scenario

name of the default scenario.

bau_scenario

name of scenario to be used for business as usual. Default is historical.

scenario_name

name of scenario

ind_ids

Named vector of indicator codes for input indicators to the Billion. Although separate indicator codes can be used than the standard, they must be supplied as a named vector where the names correspond to the output of billion_ind_codes().

...

additional parameters to be passed to scenario function

Value

data frame with acceleration_target scenario binded to df. scenario_name is set to acceleration_target

See also