Add Population Figures for HPOP Billion

add_hpop_populations() adds relevant populations to each HPOP Billion indicator and country, so these can be used to calculate indicator-level contributions to the HPOP Billion. The column specified by population will be generated and filled with relevant populations for that country and indicator. If the column already exists, only missing values will be replaced by the function.

Usage

add_hpop_populations(
  df,
  population = "population",
  pop_year = 2025,
  ind_ids = billion_ind_codes("hpop")
)

Arguments

df

Data frame in long format, where 1 row corresponds to a specific country, year, and indicator.

population

Column name of column to create with population figures.

pop_year

Year used to pull in HPOP populations, defaults to 2025.

ind_ids

Named vector of indicator codes for input indicators to the Billion. Although separate indicator codes can be used than the standard, they must be supplied as a named vector where the names correspond to the output of billion_ind_codes().

Value

Data frame in long format.

Details

For more details on the HPOP Billion calculation process and how this function ties in with the rest, see the vignette:

vignette("hpop", package = "billionaiRe")