Add scenario for a specific indicator

add_scenario_indicator gets the right add_scenario_ function for the indicator specified and adds the scenario.

Usage

add_scenario_indicator(
  df,
  scenario_function = c("aroc", "halt_rise", "percent_baseline", "linear_change",
    "linear_change_col", "quantile", "best_in_region", "fixed_target",
    "fixed_target_col", "bau", "accelerate", "sdg"),
  indicator,
  ind_ids = billion_ind_codes("all"),
  scenario = "scenario",
  ...
)

Arguments

df

Data frame in long format, where 1 row corresponds to a specific country, year, and indicator.

scenario_function

character identifier of scenario to use. Can be any of the following:

aroc: calls scenario_aroc

halt_rise: calls scenario_halt_rise

percent_baseline: calls scenario_percent_baseline

linear_change: calls scenario_linear_change

linear_change_col: calls scenario_linear_change_col

quantile: calls scenario_quantile

best_in_region: calls scenario_best_in_region

fixed_target: calls scenario_fixed_target

fixed_target_col: calls scenario_fixed_target_col

bau: calls scenario_bau

accelerate: calls indicator accelerate function.

sdg: calls Sustainable Development Goals (SDG) acceleration function.

indicator

name of indicator to be passed to scenario_function

ind_ids

Named vector of indicator codes for input indicators to the Billion. Although separate indicator codes can be used than the standard, they must be supplied as a named vector where the names correspond to the output of billion_ind_codes().

scenario

Column name of column with scenario identifiers. Useful for calculating contributions on data in long format rather than wide format.

...

additional parameters to be passed to the add_scenario_indicator function (e.g. add_scenario_adult_obese).