Recycle data between scenarios for a single scenario

recycle_data_scenario_single reuses values present in the specified scenarios in default_scenario, scenario_reported_estimated, scenario_covid_shock and scenario_reference_infilling for the specified scenarios.

Usage

recycle_data_scenario_single(
  df,
  scenario,
  billion = c("hep", "hpop", "uhc"),
  iso3 = "iso3",
  ind = "ind",
  value = "value",
  year = "year",
  type = "type",
  start_year = 2018,
  end_year = 2025,
  scenario_col = "scenario",
  default_scenario = "default",
  scenario_reported_estimated = "routine",
  scenario_covid_shock = "covid_shock",
  scenario_reference_infilling = "reference_infilling",
  include_projection = TRUE,
  recycle_campaigns = TRUE,
  ind_ids = NULL,
  trim_years = FALSE
)

Arguments

df

Data frame in long format, where 1 row corresponds to a specific country, year, and indicator.

scenario

name of scenario to recycle for.

billion

name of billion to recycle data for.

iso3

Column name of column with country ISO3 codes.

ind

Column name of column with indicator names.

value

Column name of column with indicator values.

year

Column name of column with years.

type

Column name of column with types

start_year

Base year for contribution calculation, defaults to 2018.

end_year

End year(s) for contribution calculation, defaults to 2019 to 2025.

scenario_col

Column name of column with scenario identifiers.

default_scenario

name of the default scenario.

scenario_reported_estimated

name of the reported/estimated scenario.

scenario_covid_shock

name of the scenario with the COVID-19 shock years.

scenario_reference_infilling

name of the WHO technical programs projections/imputations scenario.

include_projection

Boolean to include or not projections in recycling

recycle_campaigns

Boolean to include or not campaigns in recycling

ind_ids

Named vector of indicator codes for input indicators to the Billion. Although separate indicator codes can be used than the standard, they must be supplied as a named vector where the names correspond to the output of billion_ind_codes().

trim_years

logical to indicate if years before start_year and after end_year should be removed

Value

Data frame in long format.

Details

To do so, it looks at:

values in default_scenario but not in the scenario specified

values in scenario_reported_estimated or scenario_covid_shock but not in the scenario specified or default_scenario.

values in scenario_reference_infilling but not in the scenario specified, scenario_reported_estimated, scenario_covid_shock, or scenario_reference_infilling

For more information see:

vignette("scenarios", package = "billionaiRe")