Scenario COVID sustained disruption to normal

scenario_covid_delayed_return() creates a scenario where there is a a delayed by one year before progressivelly returning to the pre-pendemic situation.

Usage

scenario_covid_sustained_disruption(
  df,
  start_year = 2018,
  covid_year = 2020,
  recovery_year = 2022,
  progressive_recovery = TRUE,
  end_year = 2025,
  value_col = "value",
  scenario_col = "scenario",
  scenario_name = "covid_sustained_disruption",
  ind_ids = billion_ind_codes("all"),
  default_scenario = "default",
  ...
)

Arguments

df

Data frame in long format, where 1 row corresponds to a specific country, year, and indicator.

start_year

Base year for contribution calculation, defaults to 2018.

covid_year

(integer) year where the values are impacted by COVID.

recovery_year

integer year from which the AROC will be applied. Default to 2022.

progressive_recovery

(logical) TRUE if the recovery after COVID-19 should be progressive.

end_year

End year(s) for contribution calculation, defaults to 2019 to 2025.

value_col

Column name of column with indicator values. This column will be used to return the results.

scenario_col

name of scenario column to be created

scenario_name

name of scenario

ind_ids

Named vector of indicator codes for input indicators to the Billion. Although separate indicator codes can be used than the standard, they must be supplied as a named vector where the names correspond to the output of billion_ind_codes().

default_scenario

name of the default scenario.

...

additional parameters to be passed to scenario_dip_recover_iso3()

Value

a data frame with scenario values in value_col with a scenario_col column.

Details

In details, the AROC between the start_year and covid_year - 1 is applied to the last reported value to recovery_year onward in a progressive way: full AROC is only applied at end_year; otherwise the recovery is spread between the years between recovery_year and end_year. For instance, if recovery_year is 2021 and end_year 2025, then 2021 will have 0% of AROC, 2022 25%, 2023 50%, 2024 75%, and 2025 100%. If there are missing values between covid_year and recovery_year, the last value from covid_year is carried forward. This applies only to countries where the indicator value for covid_year is reported or estimated. Otherwise, the value is carried with scenario_bau.

See also