Scenario dip and recover to specific ind and iso3

Applies scenario_dip_recover() to a specific ind and iso3 combination.

Usage

scenario_dip_recover_iso3_ind(
  df,
  iso3,
  ind,
  year = "year",
  ind_col = "ind",
  iso3_col = "iso3",
  dip_year = 2020,
  recovery_year = 2021,
  progressive_recovery = FALSE,
  baseline_year = 2018,
  last_year = NULL,
  start_year = 2018,
  end_year = 2025,
  value = "value",
  scenario = "scenario",
  scenario_name = "dip_recover",
  type_col = "type",
  ind_ids = billion_ind_codes("all"),
  default_scenario = "default",
  trim = TRUE,
  small_is_best = FALSE,
  keep_better_values = FALSE,
  upper_limit = 100,
  lower_limit = 0,
  trim_years = TRUE,
  source_col = "source",
  source = sprintf("WHO DDI, %s", format(Sys.Date(), "%B %Y"))
)

Arguments

df

Data frame in long format, where 1 row corresponds to a specific country, year, and indicator.

iso3

(character) name of the ISO3 country code on which to calculate the scenario

ind

(character) name of the indicator on which to calculate the scenario

year

Column name of column with years.

ind_col

Column name of column with indicator names.

iso3_col

(character) name of column with the ISO3 country codes

dip_year

(integer) year where the dip appends

recovery_year

(integer) year from which the AROC will be applied

progressive_recovery

(logical) TRUE if the recovery after dip should be progressive.

baseline_year

(integer) identify baseline year on which the AROC should be calculated.

last_year

(integer) identify last year where values where reported or estimated between start_year and end_year.

start_year

Start year for scenario, defaults to 2018.

end_year

End year for scenario, defaults to 2025

value

Column name of column with indicator values. This column will be used to return the results.

scenario

name of scenario column to be created

scenario_name

name of scenario

type_col

Column name of column to store type.

ind_ids

Named vector of indicator codes for input indicators to the Billion. Although separate indicator codes can be used than the standard, they must be supplied as a named vector where the names correspond to the output of billion_ind_codes().

default_scenario

name of the default scenario.

trim

logical to indicate if the data should be trimmed between upper_limit and lower_limit.

small_is_best

Logical to identify if a lower value is better than a higher one (e.g. lower obesity in a positive public health outcome, so obesity rate should have small_is_best = TRUE).

keep_better_values

logical to indicate if "better" values should be kept from value if they are present. Follows the direction set in small_is_best. For instance, if small_is_best is TRUE, then value lower than col will be kept.

upper_limit

upper limit at which the indicator should be caped.

lower_limit

lower_limit limit at which the indicator should be caped.

trim_years

logical to indicate if years before start_year and after end_year should be removed

source_col

Column name of column to store source information.

source

Source to use if no unique source available for the calculation.

Value

a data frame with scenario values in value with a scenario column.

Details

In details, the AROC between the start_year and the last reported or estimated value before dip_year is applied to the last reported value to recovery_year onward. If there are missing values between dip_year and recovery_year, the last value from dip_year is carried forward. This applies only to countries where the indicator value for dip_year is reported or estimated. Otherwise, the value is carried with scenario_bau.

If progressive_recovery is TRUE, then the recovery is spread between the years between recovery_year and end_year. For instance, if recovery_year is 2021 and end_year 2025, then 2021 will have 0% of AROC, 2022 25%, 2023 50%, 2024 75%, and 2025 100%.