Trim values

This scenario allows to change a value by a fixed percentage to a provided year from a provided baseline year. It provides values for scenarios stated as "Reduce INDICATOR by XX% by YEAR"

Usage

trim_values(
  df,
  col,
  value = "value",
  year = "year",
  trim = TRUE,
  small_is_best = FALSE,
  keep_better_values = FALSE,
  upper_limit = 100,
  lower_limit = 0,
  trim_years = TRUE,
  start_year = 2018,
  end_year = 2025
)

Arguments

df

Data frame in long format, where 1 row corresponds to a specific country, year, and indicator.

col

column to trim values from. Will be removed before returning the data frame.

value

Column name of column with indicator values. This column will be used to return the results.

year

Column name of column with years.

trim

logical to indicate if the data should be trimmed between upper_limit and lower_limit.

small_is_best

Logical to identify if a lower value is better than a higher one (e.g. lower obesity in a positive public health outcome, so obesity rate should have small_is_best = TRUE).

keep_better_values

logical to indicate if "better" values should be kept from value if they are present. Follows the direction set in small_is_best. For instance, if small_is_best is TRUE, then value lower than col will be kept.

upper_limit

upper limit at which the indicator should be caped.

lower_limit

lower_limit limit at which the indicator should be caped.

trim_years

logical to indicate if years before start_year and after end_year should be removed

start_year

Start year for scenario, defaults to 2018.

end_year

End year for scenario, defaults to 2025

Value

trimed data frame, removing the col column, and putting the trimmed values in value

Details

The percent_change parameter is understood as a percentage change, and not a percentage point change, as this is usually what intended by those formulations. If it is indeed the percentage change that is required, please use scenario_aroc. For instance, to calculate the scenario "reduce the 2018 value (90%) by 30% by 2025", will results to a 2025 value of 63% and not 60%.

The returned scenario is a portion of the straight line drawn from the baseline_year value to the target_year. Only values for years between start_year and end_year will be returned.