Upload Billions indicator data

upload_billion_data() allows you to easily upload a data frame to the correct location in the World Health Data Hub. By default, it appends the current timestamp to the file name, but it can also accept a timestamp specified by the user.

Usage

upload_billion_data(
  df,
  data_type = c("wrangled_data", "projected_data", "final_data"),
  billion = c("hep", "hpop", "uhc"),
  ind_code,
  version = whdh::get_formatted_timestamp(),
  na_rm = TRUE,
  experiment = NULL,
  silent = TRUE
)

Arguments

df

(data.frame) A data frame

data_type

(string) The type of data to load.

wrangled_data (default): raw data that has been wrangled into a suitable form for analysis.

projected_data: data that has been fully projected to the target year but has not yet been transformed or calculated upon.

final_data: the complete set of billions data with transformed values, contributions, and all calculations available.

billion

(string) One of "hep", "hpop", or "uhc". Ignored when data_type = "final_data".

ind_code

(string) The name of the indicator to upload data for. Ignored when data_type = "final_data".

version

A yyyy-mm-ddTHH-MM-SS formatted string. The default is the current date time, as returned by whdh::get_formatted_timestamp().

na_rm

(logical) Specifies whether to remove rows where value is missing. Defaults to FALSE.

experiment

(string) Either NULL or a string ("unofficial" by default). Identifies where the Bronze/Silver/Gold data layers to which data is uploaded are located. Cannot be an empty string.

If NULL, the root folder for the data layers is the 3B folder (i.e., where the "official" data is stored). For example, 3B/Silver/....

If a string, the root folder for the data layers is a sub-folder within the Sandbox layer of the 3B data lake (e.g., if experiment = "my_exp", then data is download from 3B/Sandbox/my_exp/{data_layer}/...)

silent

(logical) Specifies whether to show authentication messages and a progress bar. Defaults to TRUE.

Value

A data frame. Note that this is the modified version of in the input the function (such as from removing empty rows when na_rm = TRUE or from the call to save_wrangled_output()) are carried over to the output.

Details

This function requires that the user to have the whdh package installed and setup to access the data. For quetions about getting the relevant permissions, please contact kanjim@who.int or messeillere@who.int.